adds pages, component depth, improves grid layout, centralized data in store
