+++ .gitignore
@@ -29,3 +29,4 @@ node_modules
 dist
 
 .DS_Store
+
