+++ src/components/Departure.vue
@@ -1,32 +1,39 @@
 <template>
-      <v-container>
-        <v-card class="departure-card">
-          <v-row>
-          <v-card-title>
-            <v-img
-              max-width="100px"
-              :src="operatorLogo"
-            />
-          </v-card-title>
-          <v-spacer/>
-          <v-card-title>
-            <span class="orange--text">{{priceFormatted}}</span>
-          </v-card-title>
-          </v-row>
-          <v-card-text class="departure-text">
-            <span class="time-text">{{getFormattedTime(departure.departure_time)}}</span>
-            {{getLocationAndCityNameFormatted(departure.origin_location_id)}}
-          </v-card-text>
-            <br>
-          <v-card-text class="departure-text">
-            <span class="time-text">{{getFormattedTime(departure.arrival_time)}}</span>
-            {{getLocationAndCityNameFormatted(departure.destination_location_id)}}
-          </v-card-text>
-          <v-card-text class="duration-text">
-            <span class="capitilize">{{$t('message.duration')}}</span>  {{durrationFormatted}}
-          </v-card-text>
-        </v-card>
-      </v-container>
+  <v-container>
+    <v-card class="departure-card"
+      :loading="loadingDepartures">
+      <v-row>
+        <v-col cols="6">
+        <v-card-title>
+          <v-img
+            max-width="150px"
+            :src="operatorLogo"/>
+        </v-card-title>
+        </v-col>
+        <v-col cols="6" class="d-flex flex-row-reverse">
+        <v-card-title>
+          <span class="orange--text">{{priceFormatted}}</span>
+        </v-card-title>
+        </v-col>
+      </v-row>
+      <v-card-text class="departure-text">
+        <span class="capitilize time-text">
+          {{$t('message.departure')}}: {{getFormattedTime(departure.departure_time)}} -
+        </span>
+        {{getLocationAndCityNameFormatted(departure.origin_location_id)}}
+      </v-card-text>
+        <br>
+      <v-card-text class="departure-text">
+        <span class="capitilize time-text">
+          {{$t('message.arrival')}}: {{getFormattedTime(departure.arrival_time)}} -
+        </span>
+        {{getLocationAndCityNameFormatted(departure.destination_location_id)}}
+      </v-card-text>
+      <v-card-text class="duration-text">
+        <span class="capitilize time-text">{{$t('message.duration')}}: {{durrationFormatted}}</span>
+      </v-card-text>
+    </v-card>
+  </v-container>
 </template>
 <script>
 import moment from 'moment';
@@ -36,8 +43,8 @@ export default {
   name: 'departure',
   props: {
     departure: {
-      tyoe: Object,
-      requires: true,
+      type: Object,
+      required: true,
     },
   },
   methods: {
@@ -56,13 +63,14 @@ export default {
       locations: 'locations',
       cities: 'cities',
       currency: 'currency',
+      loadingDepartures: 'loadingDepartures',
     }),
     timeLocaleFormatString() {
       switch (this.$i18n.locale) {
         case 'en':
           return 'h:mm A';
         case 'fr':
-          return 'HH:MM';
+          return 'H:MM';
         default:
           return 'h:mm A';
       }
