+++ src/store/index.js
@@ -14,6 +14,10 @@ export default new Vuex.Store({
     locations: [],
     cities: [],
     currency: 'CAD',
+    currencies: ['CAD', 'USD'],
+    languages: [{ text: 'English', value: 'en' }, { text: 'Français', value: 'fr' }],
+    cityOptions: [{ text: 'Montreal', value: 'f25dvk' }, { text: 'New York', value: 'dr5reg' }],
+    loadingDepartures: false,
   },
   mutations: {
     SET_TRIP_INFO(state, tripInfo) {
@@ -28,19 +32,19 @@ export default new Vuex.Store({
     RESET_DEPARTURES(state) {
       state.departures = [];
     },
-    SET_OPERATORS(state, operators) {
+    APPEND_OPERATORS(state, operators) {
       state.operators = state.operators.concat(operators);
     },
     RESET_OPERATORS(state) {
       state.operators = [];
     },
-    SET_LOCATIONS(state, locations) {
+    APPEND_LOCATIONS(state, locations) {
       state.locations = state.locations.concat(locations);
     },
     RESET_LOCATIONS(state) {
       state.locations = [];
     },
-    SET_CITIES(state, cities) {
+    APPEND_CITIES(state, cities) {
       state.cities = state.cities.concat(cities);
     },
     RESET_CITIES(state) {
@@ -48,6 +52,9 @@ export default new Vuex.Store({
     },
     SET_CURRENCY(state, currency) {
       state.currency = currency;
+    },
+    SET_LOADING_DEPARTURES(state, loading) {
+      state.loadingDepatures = loading;
     },
   },
   actions: {
@@ -58,11 +65,13 @@ export default new Vuex.Store({
         outboundDate,
         numberAdults,
       }) {
+      commit('SET_LOADING_DEPARTURES', true);
       dispatch('RESET_DEPARTURE_DATA');
       let index = 0;
 
       const query = `adult=${numberAdults}&currency=${state.currency}&lang=${i18n.locale}`;
-      let data = (await BusbudGateway.get(`${origin}/${destination}/${outboundDate}?${query}`)).data;
+      const pathParams = `${origin}/${destination}/${outboundDate}`;
+      let data = (await BusbudGateway.get(`${pathParams}?${query}`)).data;
 
       commit('SET_TRIP_INFO', {
         origin,
@@ -70,26 +79,31 @@ export default new Vuex.Store({
         outboundDate,
         numberAdults,
       });
-      dispatch('SET_DEPARTURE_DATA', data);
+      dispatch('APPEND_DEPARTURE_DATA', data);
+      dispatch('SORT_BY_CHEAPEST');
 
       if (!data.complete) {
         index += data.departures.length;
 
         const interval = setInterval(async () => {
-          data = (await BusbudGateway.get(`${origin}/${destination}/${outboundDate}/poll?index=${index}&${query}`)).data;
-          dispatch('SET_DEPARTURE_DATA', data);
+          data = (await BusbudGateway.get(`${pathParams}/poll?index=${index}&${query}`)).data;
+          dispatch('APPEND_DEPARTURE_DATA', data);
           dispatch('SORT_BY_CHEAPEST');
           index += data.departures.length;
-          if (data.complete) clearInterval(interval);
+          if (data.complete) {
+            clearInterval(interval);
+            commit('SET_LOADING_DEPARTURES', false);
+          }
         }, 1000);
+      } else {
+        commit('SET_LOADING_DEPARTURES', false);
       }
-      dispatch('SORT_BY_CHEAPEST');
     },
     async UPDATE_CURRENCY({ state, dispatch, commit }, currency) {
       commit('SET_CURRENCY', currency);
       await dispatch('GET_TRIPS', state.tripInfo);
     },
-    SET_DEPARTURE_DURATION({ commit }, departures) {
+    CREATE_DEPARTURE_DURATION({ commit }, departures) {
       const depWithDuration = departures.map((dep) => {
         const duration = moment.duration(
           moment(dep.arrival_time).diff(moment(dep.departure_time)),
@@ -127,11 +141,11 @@ export default new Vuex.Store({
       commit('RESET_CITIES');
       commit('RESET_LOCATIONS');
     },
-    SET_DEPARTURE_DATA({ dispatch, commit }, data) {
-      dispatch('SET_DEPARTURE_DURATION', data.departures);
-      commit('SET_OPERATORS', data.operators);
-      commit('SET_LOCATIONS', data.locations);
-      commit('SET_CITIES', data.cities);
+    APPEND_DEPARTURE_DATA({ dispatch, commit }, data) {
+      dispatch('CREATE_DEPARTURE_DURATION', data.departures);
+      commit('APPEND_OPERATORS', data.operators);
+      commit('APPEND_LOCATIONS', data.locations);
+      commit('APPEND_CITIES', data.cities);
     },
   },
   getters: {
@@ -139,6 +153,10 @@ export default new Vuex.Store({
     operators: state => state.operators,
     locations: state => state.locations,
     cities: state => state.cities,
+    currencies: state => state.currencies,
     currency: state => state.currency,
+    languages: state => state.languages,
+    cityOptions: state => state.cityOptions,
+    loadingDepartures: state => state.loadingDepartures,
   },
 });
